update the package name.
