+++ impl/src/main/java/com/starrocks/connector/kettle/steps/starrockskettleconnector/StarRocksOP.java
@@ -1,0 +1,30 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.starrocks.connector.kettle.steps.starrockskettleconnector;
+
+public enum StarRocksOP {
+    UPSERT, DELETE;
+    public static final String COLUMN_KEY="__op";
+
+    static StarRocksOP parse(String option){
+        if (option.equals("UPSERT")){
+            return UPSERT;
+        }
+        if (option.equals("DELETE")){
+            return DELETE;
+        }
+        throw new RuntimeException("Unsupported row option.");
+    }
+}
