+++ impl/src/main/java/com/starrocks/connector/kettle/steps/starrockskettleconnector/starrocks/StarRocksJdbcConnectionOptions.java
@@ -1,0 +1,57 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.starrocks.connector.kettle.steps.starrockskettleconnector.starrocks;
+
+import javax.annotation.Nullable;
+import java.util.Optional;
+
+public class StarRocksJdbcConnectionOptions {
+
+    private final String url;
+    private final String driverName;
+    private final String cjDriverName;
+    @Nullable
+    private final String username;
+    @Nullable
+    private final String password;
+
+    public StarRocksJdbcConnectionOptions(String url, String username, String password) {
+        this.url = url;
+        this.driverName = "com.mysql.jdbc.Driver";
+        this.cjDriverName = "com.mysql.cj.jdbc.Driver";
+        this.username = username;
+        this.password = password;
+    }
+
+    public String getDbURL() {
+        return url;
+    }
+
+    public String getCjDriverName() {
+        return cjDriverName;
+    }
+
+    public String getDriverName() {
+        return driverName;
+    }
+
+    public Optional<String> getUsername() {
+        return Optional.ofNullable(username);
+    }
+
+    public Optional<String> getPassword() {
+        return Optional.ofNullable(password);
+    }
+}
