+++ impl/src/main/resources/com/starrocks/connector/kettle/steps/starrockskettleconnector/messages/messages_zh_CN.properties
