+++ ui/src/main/resources/com/starrocks/connector/kettle/ui/starrockskettleconnector/messages/messages_zh_CN.properties
